// src/interfaces/color.ts
export interface ColorProps {
  navy: string;
  navyLight: string;
  navyDark: string;
  ocean: string;
  oceanLight: string;
  oceanDark: string;
  grassGreen: string;
  grassGreenLight: string;
  grassGreenDark: string;
  tangerine: string;
  tangerineLight: string;
  tangerineDark: string;
  greyCool: string;
  greyCloudy: string;
  greyCharcoal: string;
}